backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
openai==1.3.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
fal-client==0.1.0 